Added support for selective linkage of port drivers, including nvs, network, socket, gpio, i2c, spi, ledc, and any other port drivers in the future.  By default, all drivers are added in the future.
