+++ .gitignore
@@ -6,6 +6,8 @@ xcode/**
 src/platforms/esp32/build/**
 src/platforms/esp32/build/**/*.d
 src/platforms/esp32/components/**
+src/platforms/esp32/main/component_nifs.txt
+src/platforms/esp32/main/component_ports.txt
 !src/platforms/esp32/components/libatomvm/
 .idea/**
 .vscode/**
