+++ src/platforms/esp32/main/nvs_nif.h
@@ -1,0 +1,24 @@
+//
+// Copyright (c) dushin.net
+// All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
+#ifndef _NVS_NIF_H_
+#define _NVS_NIF_H_
+
+void nvs_nif_init(GlobalContext *global);
+const struct Nif *nvs_nif_get_nif(const char *nifname);
+
+#endif
